trying to use less bits in key field IDs

+++ src/fields.go
@@ -1,9 +1,9 @@
 package edb
 
-type IntArr map[int]IntField;
-type StrArr map[int]StrField;
-type SetArr map[int]SetField;
+type IntArr map[int16]IntField;
+type StrArr map[int16]StrField;
+type SetArr map[int16]SetField;
 
 type IntField int32
 type StrField int32
-type SetField []int
+type SetField []int32
